+++ src/ECS/Resources/FBO_EnvMap.h
@@ -36,7 +36,7 @@ struct FBO_EnvMap : FBO_Base {
 		m_size = glm::ivec2(width, height);
 		for (int x = 0; x < 6; ++x) {
 			const glm::ivec2 size(glm::floor(glm::vec2(m_size) / glm::vec2(pow(2, x))));
-			glTextureImage3DEXT(m_textureID, GL_TEXTURE_CUBE_MAP_ARRAY, x, GL_RGB8, size.x, size.y, depth, 0, GL_RGB, GL_FLOAT, NULL);
+			glTextureImage3DEXT(m_textureID, GL_TEXTURE_CUBE_MAP_ARRAY, x, GL_RGB16F, size.x, size.y, depth, 0, GL_RGB, GL_FLOAT, NULL);
 		}
 	}
 	void clear(const GLint & zOffset) {
